add instrument=-1. Fix a bug of float speed, trackID out of bounds, not auto switch to output text panel

+++ src/main/java/interpreter/entity/Paragraph.java
@@ -14,7 +14,7 @@ public class Paragraph {
 
     private byte volume;
 
-    private byte instrument;
+    private int instrument;
 
     private final List<Integer> noteList;
 
@@ -42,11 +42,11 @@ public class Paragraph {
         this.volume = volume;
     }
 
-    public byte getInstrument() {
+    public int getInstrument() {
         return instrument;
     }
 
-    public void setInstrument(byte instrument) {
+    public void setInstrument(int instrument) {
         this.instrument = instrument;
     }
